Change []runes to be an interface, Runes.
What
===
Change []runes, that's used as the source of runes to use for the
randomly generated string to be an interface and added RuneArray which
implements Runes.

Why
===
So that the source of the runes does not need to be a prefilled array
and could be in the future other sources like a range, stream, etc.