Add the date to the version number.
What
===
Add the date to the version number.

Why
===
So that the version number is always incrementing. Since the version
number was previously only the short git SHA1, there was no guarantee it
would be increasing with each new version and Homebrew will only install
a new version if the version number is greater than the last.